% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{compress_data}
\alias{compress_data}
\title{Compresses a dataset according to new coordinates}
\usage{
compress_data(df, U)
}
\arguments{
\item{df}{A matrix. The original dataframe. Each row corresponds to an observation (n rows, p columns)}

\item{U}{A matrix. The new coordinates (PC). Each column corresponds to a PC (p rows, k columns)}
}
\value{
A matrix with n rows and k columns.
}
\description{
Compresses a dataset according to new coordinates
}
