% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{fraction_variance_explained_perPC}
\alias{fraction_variance_explained_perPC}
\title{Fraction of variance explained per PC}
\usage{
fraction_variance_explained_perPC(C, U)
}
\arguments{
\item{C}{A matrix. The correlation or covariance matrix (p x p).}

\item{U}{A matrix. The matrix containing the r PCs (p x r).}
}
\value{
An array.
}
\description{
Computes the fraction of variance explained (variance explained normalized by the trace of the covariance/correlation matrix) by each PC.
}
