% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{orthogonality_violation}
\alias{orthogonality_violation}
\title{Computes the orthogonality constraint violation defined as the distance (L infinity norm) between $U^\top U$ and the identity matrix}
\usage{
orthogonality_violation(U)
}
\arguments{
\item{U}{A matrix. Each column correspond to an n-dimensional PC.}
}
\value{
A float.
}
\description{
Computes the orthogonality constraint violation defined as the distance (L infinity norm) between $U^\top U$ and the identity matrix
}
