% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{variance_explained_perPC}
\alias{variance_explained_perPC}
\title{Variance Explained Per PC}
\usage{
variance_explained_perPC(C, U)
}
\arguments{
\item{C}{A matrix. The correlation or covariance matrix (p x p).}

\item{U}{A matrix. The matrix containing the r PCs (p x r).}
}
\value{
A numeric vector of length `ncol(U)`.
}
\description{
Computes the variance explained by each PC.
}
