% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{fraction_variance_explained}
\alias{fraction_variance_explained}
\title{Compute the fraction of variance explained (variance explained normalized by the trace of the covariance/correlation matrix)}
\usage{
fraction_variance_explained(C, U)
}
\arguments{
\item{U}{A matrix. The matrix containing the PCs (p x k).}

\item{Sigma}{A matrix. The correlation or covariance matrix (p x p).}
}
\value{
A float.
}
\description{
Compute the fraction of variance explained (variance explained normalized by the trace of the covariance/correlation matrix)
}
