% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{orthogonality_violation}
\alias{orthogonality_violation}
\title{Computes the orthogonality constraint violation defined as the distance (infinity norm) between $U^T U$ and the identity matrix}
\usage{
orthogonality_violation(U)
}
\arguments{
\item{U}{A matrix. Each column correspond to an n-dimensional PC.}
}
\value{
A float.
#' @examples
library(datasets)
TestMat <- cor(datasets::mtcars)
mspcares <- mspca(TestMat, 2, c(4,4))
orthogonality_violation(mspcares$x_best)
}
\description{
Computes the orthogonality constraint violation defined as the distance (infinity norm) between $U^T U$ and the identity matrix
}
