% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{print.mspca}
\alias{print.mspca}
\title{Displays the output of the msPCA algorithm}
\usage{
\method{print}{mspca}(sol_object, C)
}
\arguments{
\item{sol_object}{A list. The output of the mspca or twp function.}

\item{C}{A matrix. The correlation or covariance matrix (p x p).}
}
\description{
Displays the output of the msPCA algorithm
}
\examples{
library(datasets)
TestMat <- cor(datasets::mtcars)
mspcares <- mspca(TestMat, 2, c(4,4))
print.mspca(mspcares, TestMat)
}
